+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added formatDate util
 - Added RowCard component
-- Added contact section
+- Added contact section, linked resume
 - Updated navigation to its own component
 - Updated work and footer sections
 - Updated README